Initial commit
This is a combination of the following crates, all from Stjepan Glavina:

- async_task
- async_io
- async_executor
- multitask

Executor:
---------

The code is adapted to allow for a thread-per-core framework to be
implemented. Atomics are removed almost everywhere, and support
is added for multi-queue in-thread execution.

Comments:

I feel like the state machine in raw.rs can be greatly simplified
with the TPC assumption, but that would need a lot of stabilization
effort, so for now we will leave it as is. I fixed some issues during
the implementation of yield (double scheduling), but that's good enough
for now.

I/O
---
The epoll&friends mechanism is replaced with io_uring, which makes
this Linux-specific.

Comments:

More stabilization work is needed during shutdown. I am sometimes seeing
issues with de-registering the buffers with indicate drops may be going
awry. Speaking of buffers, they are there but we are not using it yet.

It will have to come later.