bring back the atomics in task implementation (temporary)
I have been suffering from random memory corruptions during task allocation /
deallocation. Also having a hard time getting the sanitizers to work reliably
in the nightly compiler.

So temporarily, bring the atomics back so we can keep making forward progress.

The code is now essentially what is in async-task, with the exception that the
schedule function is marked as !Send + !Sync (while async-task's is Send + Sync),
and the crate use paths are different.

Hopefully we can live with that for a long time but if not we'll deal with that
later.

Those memory corruptions were also the cause as to why the tests with join_all
were failing, so we will now bring back the good code (next patch)