do not pass a timeout for sleep
It doesn't mean what we think it means: this is not an uring primitive,
but rather an iou construct. It will register a timeout event, and set
-1 to its user data. Because we trust the user data to have a Source
object this can never, ever be used by us.

And we can safely remove it: if it was being used, we would have
crashed by now.

We will soon have our own timeouts that will work well with our
Source objects.