drop uring buffers for now
I am seeing annoying messages during tests that buffer de-registration
failed. This is because buffer registration is currently commented out:
if they are not registered, they cannot be deregistered.

To avoid clutter, we will for now just do away with the code. It is a bit
bitrot at the moment anyway because some methods were added to the posix
buffer implementation that are not present in the uring buffers implementation.