remove atomic tick increment
This tick was placed here when this was part of async_io and was used
by the epoll mechanism. In the uring implementation I never used it
and removed most, but left this one behind.

It is a good thing to have a poll tick counter -- seastar has one that
is not atomic to expose metrics but since we don't yet have integration
with any metric framework we will remove for now.