consume returned cqes when we are forcing a submission
Force submission happens when we are removing a timer, since we need
to make sure that the completion for the cancellation won't be posted
together with the new registration.

io_uring can return EBUSY if the cqe ring is full, so we did that on
a loop. However because we didn't consume the returning cqes, if that
failed once it would fail forever.

To make this explicit I am separating the submission code into its own
method. Because it can generate events, wakers has to be passed.