fix runnable documentation examples
Cargo compiles and runs code snippets in code documentation when testing
code. This commit updates them so that they are valid and use the latest
APIs.

Because scipio is a TPC async framework, it is sometimes difficult to
demonstrate a feature in the documentation (for example, the `Parker`
object). In these cases, they are marked as `no_run`.

Some other code snippets were left as-is because they rely on APIs that
don't exist anymore. For example, `multitask` examples that rely on a
now removed `LocalExecutor::ticker`.