Always have a preempt_timer
Our strategy of not having a preempt timer when there are no latency
classes is a tad too aggressive: it assumes that the queues will
eventually be exhausted by the throughput classes, but some classes
can generate infinite work through inifite loops. Provided they are
well behaved enough to call need_preempt(), we want to eventually
break them out of those loops.

One of our tests is explicitly testing that in the absence of
latency tasks we don't get preempted for a second. And because we don't
want the preempt timer to take a lot longer than that in practice
if we are to have one, change the test to test for 200ms.