do not mangle timer expiration and timeout together
Because we are using the timeout parameter to park as the preempt_timer,
it can't be conflated with the timer timeout.

For instance, if we have a timer registered to fired 1s in the future,
we can and should still sleep, as long as we make sure we wake up at
the right time.