test that the fallback drop code is working
We don't want to be left with files unclosed but can't support async drop.
So users are expected to drop their files asynchronously, but if they don't
we'll do it synchronously

Ivan (@sitano) recently brought into question whether or not this was
really working so it is useful to come up with a unit test for that.

The good news is that it is. But Ivan is right that the drop statement
isn't doing much since it would naturally happen anyway and it is just
dropping the reference.

This patch removes the drop() - as Ivan suggested, but doesn't
replace with an explicit close as it is not needed.

And now that we have a unit test this should keep working forever.