correctly handle partial reads
Hippo's unit test surfaced the fact that we are not correctly
dealing with partial reads. We need to check the value returned
by read, and if necessary adjust the size of the buffer.

For the aligned read version we can use the result of the operation
directly.

For the non-aligned, it might have been that the application asked
for a size that got padded, so we need the minimum between them.