remember if a file is pollable
When creating or opening a file, we remember if the file supports
polling io based on `O_DIRECT` supports.
This detection is performed when a file is first created or opened.

For now we support two configurations:
* either the fd can be open with `O_DIRECT` and we assume it is pollable
* or it is not pollable at all and will use the sleepable ring

This excludes some categories of devices like HDDs (compatible with
`O_DIRECT` but not with polling). For now (and most probably, ever) we
ignore this case.