unit test both tmpfs and (optionally) compatible storage medias
Run the units on both tmpfs and, optionally, on a user-provided
directory that supports poll mode io_uring interactions. If none is
provided via the `SCIPIO_TEST_POLLIO_ROOTDIR` environment variable, a
message is printed on stderr, but the test don't fail.

I modified all tests to run twice and included a mechanism for them to
change their assertions depending on the current target directory. This
allows `file_allocate` to expect success for non `IO_DIRECT` files and
expect a failure otherwise.

I am not certain this is the right approach though. It may be simpler to
just duplicate all tests in practice.