update parking documentation and related cleanups
The parking documentation is outdated. Update it to match what we are doing.
Also finally remove the Unparker (which was a nop before already) to reflect
the fact that the documentation now says we don't have one.

This is not supposed to be used by other crates, so all previously public
functions are moved to be pub(crate). As we do that, examples that relied
on the Parker being public fail to compile. I see little value in keeping
them around because they are fairly obvious so I just removed them.