import iou, uring-sys
Instead of using git pointers on Cargo.toml, import this code directly.
Unlike the code in async-executor et al, which is heavily modified, I
intend this to be temporary until all the changes we need in iou and
uring-sys are pushed upstream.

From iou, we need this merged (not merged as of today):

  https://github.com/withoutboats/iou/pull/46

From uring-sys, we need a version that has this patch from liburing
*reverted*:

  f0c5c54945ae92a00cdbb43bdf3abaeab6bd3a23

liburing's repo is correct, but when I tried to just move to a newer
HEAD on uring-sys side I had issues because the cqe structure changed
sizes in liburing vs its representation in uring-sys. Lost a day on
this =(

Regardless of revert or move forward, that needs to be in a published
version of uring-sys so we can publish to crates anyway. Might take
time.

Aside from that, changes here are:

* I ran rustmft just in case
* pub functions become pub(crate). We then need to allow dead code
  and dead imports to avoid making a mess
* mod statements are made use statements
* the resultify macro is copied to its users because importing
  macros without touching the main crate file is a challenge.

The goal is to keep changes to the actual Scipio code to a minimum,
so we can easily swap that later.