ergonomics: provide Local, a nicer way to write Task::<()>::
Some functions that access the executor are available through Task,
which allow us to access the local executor without having a reference
to it (through its local thread local variable).

However Task's main role is to spawn a new task and store its result.
So it requires a type generic parameter, T.

For helper functions that don't spawn we have to write Task::<()>::
which is a hassle.

This patch adds Local, an alias to Task::<()>::