spawn_new: an ergonomic way to spawn a new executor.
It is easy to spawn an executor in the current thread: all we have to do is
create it.

However, there are two main problems with it:

 1. The future that we create initially does not live in a task queue. This is
    pretty much a requirement, because the run() function just executes a future
    to completion. There are real problems arising from the fact that this future
    is not on a task queue. For example, we can't call later().await on it (as
    what that does is move ourselves to the back of the task queue).

 2. The potential of a thread-per-core system is only realized when there are
    many executors, one per each CPU. That means that the user now has to create
    thread, keep them alive, create an executor inside each thread, etc. Not fun.

To make this code more ergonomic, we will create the function
LocalExecutor::spawn_new().  It will create a thread and spawn an executor
inside it, making the creation of executors easier.

The future that the caller passed is executed inside a task queue, meaning
every Scipio function will work.  And he future that we actually pass to run,
is the result of draining that task queue.

When the user is done, the function wait_on_executors() can be called to wait
on all the executors created this way at once.

Like so:

LocalExecutor::spawn_new("hello", Some(0), async move {
    function_on_cpu0().await;
});

for i in 1..N {
	LocalExecutor::spawn_new("hello", Some(i), async move {
	    function_on_cpu_aux(i).await;
	});
}

LocalExecutor::wait_on_executors();

The next step in ergonomics for this is having a single entry point
that creates N executors and wait on them. However it is a bit hard
to figure out the details of how this should look like without having
inter-executor channels for communications. So we will defer.