executor: pass a function rather than a future
The main difference in usage is that now instead of writing
"async move" we write "|| async move"

This is because we can't really require this future to be Send, or
it starts to requiring everything down the line to be Send too.

But we still don't want to just lift the requirement. So we require
FnOnce + Send.