return a JoinHandle on detach
The way detach works right now falls a bit short of my expectations.
The function takes a self parameter (not a reference to self), which
causes it to consume the task. From this moment on it is impossible
to obtain a handle. Rust move rules would have prevented us from
taking ownership of the handle before detach() is called too. One
nasty side effect of this is that once you detach, there is no nice
way to cancel.

This patch changes it slightly: detach() now returns the handle it
consumed. A user that captures that handle can now call .cancel()
on it.