introduce TimerAction, an ergonomic way to deal with timers
Timers block the current task queue making they cumbersome to use:
to use a timer one has to create a task with the timer on it, spawn
it, then decide whether to detach or wait. All doable, but all more
work than necessary.

As we strive to provide nice, ergonomic APIs in Scipio, this patch
introduces the TimerAction. That is a convenient way to execute
an action once or periodically in the future.