provide a function to grab the current task_queue
We have Local::local() and Local::local_into(), the former launching
a task in the current task queue and the latter into a specific task
queue.

That is enough when the user is in control over which to pick, but
there are situations where we want to implement helper functions
and want x and x_into variants for the helper functions as well.

In such situations, we could reduce code duplication if we knew
which task queue we are currently running.

Then instead of writing Local::local(),
we can write Local::local_into({}, current_task_queue()).