timers: keep their id stable
I am about to add a patch that allows us to cancel TimerActions without
spawning new tasks. To do that, it really helps if the timers have some stable
key that can be used to find and cancel them in the reactor.

Right now we don't: as a heritage of the old async-io threaded design, timers
are put in a queue (that does atomic operations), then the queue is processed
when we are at the correct thread. Adding to that queue creates and ID, which
is different for every timer that is a property of the operation in the queue,
not the timer.

We have no need for such queue because we are always in the right thread.
Removing this queue:
 - remove some more legacy atomic operations
 - allow us to easily extend and simplify the API to have stable IDs for each
   timer.

To avoid races we never reuse IDs. They come from a monotonic pool.