timer_action: provide a destroy() method
cancel() consumes the task and is an async method. Hippo had already
noticed how it would be nice not to do so, and I just found an example
in a real application where it became a real pain:

If you are at a point where you can't block on an .await, the only way
to cancel the TimerAction is by spawning.

This is problematic for a variety of reasons:
 - Under load, it could be that our task queues become big. We may not
   be able to cancel the timer until it had already fired.
 - By adding a new task to the queue we are contributing to the load,
   which may lead to trashing if the load is mostly due to timers.

Not to mention now we are left with a task that has to be either waited for
(and if we could wait we wouldn't have this problem) or detached. If we
have too many detached tasks, our task queue may blow up.

For this reason we'll provide a destroy() method.

We still want to keep cancel() as is for consistency with other parts of
the library (the task's cancel() is async). But we will reimplement cancel()
in terms of our newly added destroy.

There is a unit test added that guarantees that we can join a destroyed task.