split TimerAction in two
A TimerAction can now be created as a result of a one-shot event, or
run repeatedly.

As I add more functionality it is starting to feel like those two
things are a bit overloaded and some actions may not make sense for
both. In the next patch I will introduce a way to reprogram a one-shot
action, which definitely doesn't make sense for a repeat action.

The fact that by splitting them we are able to gracefully get rid of
the ugly "DurationLike" helper struct is an indication that this is the
right direction: as the TimerAction(Repeat) always have to return a
Duration, it should not depend on a external type T!

A new test is added to make sure that the return from join for the
Repeat action is what we expect.