add an inner member to Source
Source has been returning a Pinned Box directly to users. When users are done
with it, the box will be freed which will free its memory.

That's horrible, because references to that source may still be in-flight in
the io_uring. I would like to change that to a reference counted object, but to
do so we need to use an inner pattern.

If we return the reference counted object directly, the task will never die
until it goes through the ring. We don't want that: we want to allow the task
to die (for instance, by being explicitly cancelled), and want that event to
allow us to trigger a cancellation event.

So when the Source dies, we will have an opportunity to trigger a cancellation
while the inner part of the object is reference counted and still has the
resources alive.

To make the transition easier this patch does not deal with the reference
count. Only the public interfaces are changed so the path forward becomes
easier.