allow requests to be properly cancelled
Cancellation was never properly implemented. Shame on me: but at least I left a
panic around so I would get to it eventually.

Rafael hit such panic in issue #52, so it's time for us to fix it.

Turns out that the reason I didn't implement it at first is that it is quite
hard. It is also quite hard to test and find a test that reliably hits the
right cases.

Rafael provided such test.

In previous patches the Source was converted to use an inner pattern, and that
inner pattern is a Rc. We will keep a list of registered sources.  They will be
registered throughout the entire life of the request. So for as long as the
request is alive in the io_uring, it will keep the reference count up. When
the request completes or is cancelled, we get rid of it.

Two test cases that stress cancellations are added, inspired by what Rafael
provided.