shared: a cheaper version of Rc<RefCell<>>
Daniel has been working to add linters to make sure that a borrow() on a
RefCell never crosses a suspension point.

Still, I've been thinking of ways to not pay the cost of the runtime check.

This proposal implements such a way: by allowing scoped access to the interior
of an Rc we should be able to guarantee that accesses are safe.

That is done through a new struct, Shared, which resembles an Rc. However
it will not implement Deref, meaning we can't get a reference to it outside
the scope of a closure passed to it.

The closure that we expect is not async, meaning we cannot have a suspension
point inside it, and because all spawn-like functions expect a 'static lifetime
and the reference is only valid inside the closure we can't spawn either.