split the submission queue in two
We'll now have two queues: the submission queue and the cancellation
queue. The cancellation queue has to be flushed immediately. It is racy
to leave cancellation requests dangling as they may refer to resources
that were freed and are now reused.

The queues are also made Rc<RefCell, so we can tell the source about who
originated the requests.