introduce `LinkStatus` to mark a ring as "linked" to another
The commit introduces an enum type instead of a `bool` to flag whether a
`SourceType::LinkRings`'s fd is linked or free-standing.

Scipio's single-threaded reactor can sleep on a ring until an event is
available. Because we have more than one ring, we need a way to somehow
sleep on multiple rings at once. This is done by "linking" a ring to
another, i.e adding (ala poll_add) another ring's fd in the ring the
reactor will sleep on before going to sleep. In effect, this allows the
reactor to be woken up if either of the two rings have new events ready.