waker_fn : silence clippy
What this function is doing is bumping a reference count in a Rc field
that will be asynchronously freed later by a call to drop(). Therefore
we want to clone the value but won't really use it. We just want to see
the refcount go up.

Clippy doesn't like that, because in normal code that means someone just
issued an unnecessary clone.

There might be ways to write valid rust code that doesn't trigger clippy
by adding helper dummy functions and whatnot, but this code is already
complex and sensitive enough as is.

So the right thing to do is just to manually silence clippy, adding
a comment as to why we are doing it.