multitask: do not Rc + Box
We already have an Rc around the callback, so we don't need to Box.  Clippy
doesn't quite like it.

The main issue is that because we are dealing with trait pointers (Fat
pointers), we can't compare them directly either since the vtables may be
different.

There is a way to fix that comparison if we can guarantee that the callbacks
are always craeted through safe use of Rc::clone(): If they are, that means
that the data address should be the same.  We can convert the pointer to a data
pointer before comparing, and this should give us what we want.

However the equality code is not currently used right now by any data structure
and only sits here for completion. It's safer to delete it, and we'll cross
that bridge when we come to need it for real