source: move most of the source code to the io_uring file.
And make its methods fn instead of pub(crate). The code before
lived in mod.rs, closer to its definition point. However it does
some nasty unsafe things to account for the fact that io_uring is
an asynchronous interface, and we're associating a Source to each
request.

The functions that return references to core parts of Source are all internal
now, and the ones that are pub(crate) are all returning copies of parts of the
source the external entities need to manipulate (type, raw fd, etc)

Also add a comment about why the stuff clippy doesn't like is fine