uring: correctly set up timer before going to sleep
There are two problems plaguing us today when registering a timer before
going to sleep:
1. we are not flushing the latency ring, as rearm_preempt_timer relies
   on that happening outside the caller.

2. We are not consuming the cancellation for the existing preempt timer

Because of 1, the timer is not really installed. But because of 2,
we wake up anyway to consume the cancellation of the preempt timer.
This turns a bug that would be easy to detect (timers never firing)
into a more subtle bug: CPU usage much higher than it should be if
there is a timer registered.

I believe that this was facilitated by the fact that all timer handling
was stashed together in a single function, rearm_preempt_timer. This made
reasoning about it a bit harder.

To fix that I am separating the cancel and the rearm part of
rearm_preempt_timer in two and making them return the empty type. This
makes it obvious that the rings have to be flushed.

The last piece of the puzzle is that flush_cancellations have to be
changed to not only put the cancellations in the ring but also consume
them.

Kudos to Ivan for finding this issue.