fix poll ring not returning any events
In commit af8b7159f48a01fa6a1e54c0e3381656eb1a0564, where the submission
queue was split in two, code was added that would only call io_uring_enter
if requests were really generated from the queue. That was done because
otherwise now that we have twice as many queues, the cost of flush would
have been twice as high. The cancellation queues are almost always empty
so we wanted to protect against that.

However that logic doesn't work for the poll ring: if there are requests
pending in the poll ring, we must always enter.