file stream: convenience API for sequential access to files - writes
Accessing files randomly is not always what we want. As a matter of
fact, for Writes sequential access is the norm.

Rust's AsyncWrite functions are sequential access traits. As we
add our own sequential access structs, we will implement those traits
to make sure that we play well with the broader community.