rename file by names
rename() is a syscall that takes two file paths. While it may be convenient to
be able to rename the current file, there are situations in which it is not:
for instance, if we have a file stream, where there is no direct access to the
underlying file.

While we could somehow implement rename for the file stream, it is more general
to just provide a rename version that matches what Linux would expect.