implement two new operations for the StreamWriter
current_flushed_pos() will allow us to ask the writer how much data
is already flushed and made into the disk. New readers are guaranteed
to see this data except if there is a crash.

sync() is a crash-safe sync operation (in this implementation fdatasync)
that will be guaranteed to flush the data all the way down to the
underlying storage, and return which position is already safely stored.