facility to allow testing if two files are the same
file1 == file2 tests for object equality: it will return true if
the two files are the same file descriptors. However it is possible for:

- hard links to exist
- open a file more than once.

In which case they will represent the same on-storage entity but will be
represented by two different objects. This patch adds a facility that
allows us to test for that.