need_preempt: allow users to use need_preempt() directly
We have been enclosing need_preept() into a helper yield_if_needed()
that returns a future, allowing the user the convenience of .awaiting
directly on it.

However there are situations where we can't really .await, for example
if we have a borrow. For those, export need_preempt() directly. It
comes with an example of how to use each in the examples directory.