Fix bug in sync_open
`std::path::Path` doesn't include trailing \0, we need to convert it
to CString first.

I wonder if it makes sense to use stack-allocated buffers for smaller
paths here, string allocation for every syscall seems excessive. On
the other hand, this is exactly what stdlib is doing, so maybe this is OK?