add tests for ErrorEnhancer, and fix into_inner()
ErrorEnhancer doesn't have its own tests. Instead it is relying on
indirect unit tests in other parts of the crate that deals with errors.

And guess what? It broke.

This patch adds unit tests specifically targetted at making sure that
the ErrorEnhancer is working and doing what it is supposed to do, and
relaxes the error condition in the other files that were checking for
specific error strings.