remove unnecessary use of unsafe in buffers
After Aleksey's good work, there are still two usages of unsafe inside
DmaBuffer that could be avoided.

Aleksey suggests we could probably just use the standard function
copy_from_slice and do away with our version.

However our version is very convenient in the sense that it doesn't
require the slices to be of the same size and does size management
for us.

What we can do, is replace the uses of unsafe with copy_from slice.

The problem is that now the name is confusing because we have two
things with the same name that behave differently. So rename them
read_at and write_at. The _at suffix makes it clear that we are
dealing with an offset.