executor: change way we calculate shares
This will make it slightly easier to implement controllers: shares
are not either static or dynamic. Static shares work similarly to
a const variable (we can optimize because we know they never change)
whereas dynamic shares may change with time.

The methods to set and get shares are gone from visible API. Tests
are also added to make sure that the scheduler is working as intended.

While we are at it, document the whole shares mechanism better as
it is currently underdocumented.