Implement our first channel
Channels are useful abstraction when data needs to be passed between
asynchronous entities.

I have just came across a use case where channels would provide a much
more ergonomic way to code than the Deque, so here's our first!

The LocalChannel is an executor-local channel (meaning !Send, !Sync)
that is useful to pass data between task queues.

The use case for this is an internal service that needs to pick up
work to do serially. Imagine for example a flush service that wants
to flush one file at a time.

The flushers will live in a separate task queue. The entities generating
the work now have to register it into the flusher's task queue.

Using a Deque is possible but you now have to wrap it under an Rc
and code a loop-like construct (and hey, if that's what float your
boat, have at it!)

Using the LocalChannel, however, we can write this:

  let (sender, receiver) = LocalChannel::new_unbounded();

  Local::local_into(async move {
    receiver.for_each(|x| { do_something(x) }).await;
  }, tq).unwrap().detach();

  sender.send(...);