fix issue with yielding queues
yield is not working as we intended. When we yield we put ourselves at
the back of the line for our task queue, but we if there are other tasks
in our task queue we keep executing them until `need_preempt`. This is
not how this was intended to work: yield is supposed to make us go execute
another task queue.

This change tests that, while adding a test so we don't rebreak it in the future