increase scheduler resolution
Scheduler is dealing with runtime calculations in microseconds.
Although in practice this is likely enough, some of our tests are
starting to show that for very aggressive applications this may not
be enough (it is a lot of overhead to actually measure all this and
you should never have such small tasks anyway, but that's a different
story)

If a task group does run for less than 1us the scheduler goes completely
crazy, as it may accounts 0 as its vruntime. Since we plan to be general
purpose, I'd prefer at this time to pay for the extra cost of dealing with
nanos then having the scheduler misbehave.

The convertion to nanos in itself is not the issue, but we now have to
check for overflow since accounting in nanos would make us run out of
vruntime in around a year. It's a branch that should not give the predictor
any trouble, so we'll pay it.