Fix unsoundness in DmaFile and Directory
`-1` is not a valid file descriptor, and Rust can actually exploit
this fact for niche filling, so a `-1` can cause UB (in theory, if not
in practice yet).

So, instead of overwriting fd with a dummy one, let's just *forget* to
drop the thing. Ideally, we'd only forget the `file` itself, and not
the whole struct, but that requires destructing a type which
implements `Drop`, which is not possible.

So let's just manually drop heap-allocated things. This could be
mildly annoying to keep in sync with the list of fields, but probably
less so than dealing with a dummy `Option<std::file::File>`.

If this pattern become pervasive, we could use

https://crates.io/crates/derive_destructure

to automate it, but, as it is a proc macro, let's avoid introducing
that dependency for now.

closes #92