Fix use after free during cancellation
`InnerSource` is a resource shared between the kernel (uring) and user
application.

Just after submitting the op, `InnerSource` has two owner (rc == 2):

* the `Source`
* the `SourceMap`

Without this patch, if the user drops the `Source` before `.await`ing
its completion, both owners relinquish their ownership. The `Source`
is being dropped, and we `map.map.remove(&id)` from the `SourceMap`.

This causes `InnerSource` to be dropped, *while the kernel still uses
its buffers*. That is, the cancellation is too eager -- we free
resources of the op while the op is still in flight.

The suggested fix is to make freeing the resources lazy. Specifically,
we only remove the entry from SourceMap after the corresponding op
completes. Unless, of course, we haven't submitted the op in the first
place -- in this case, eager cancellation is valid.

As a side effect, `consume_source` and `add_source` now need to match
exactly.