standardize on futures_lite
This patch removes "futures" from Cargo.toml, meaning it won't be possible
anymore to use it by accident in our implementations.

There are still a lot of users in test code, so for now we'll leave it
in dev-dependencies so we don't need to rush to convert.

Outside of testing, really the only contentious use was in file_stream.