Fix use after free with using timers
We give the kernel a *pointer* to a timespec struct. Keeping this
struct on the stack leads to a textbook use-after-free.

Instead, lets stash timespec into Source, the same thing we do for all
other resources shared with the kernel.

We also introduce a helper `TimeSpec64` struct for more idiomatic
API (notably to make `derive(Debug)` work on Source).