implement async reading from stdin
Now that we have support for stream reading files that are not Direct I/O
we can implement asynchronous versions of stdin

For the moment, stdout and stderr are not implemented: I don't find them
terribly useful at the moment. Sure there is an internal lock, but high
performance operations are usually done through log files, not stdio/ stderr.
Most importantly, the act of writing is synchronous to a in-memory file.

Stdin is still useful because the read operation is usually a blocking pattern
like

 stdin().read_lines.await.