fix broken links
I have just used linkchecker (we should add that to CI) to check for
broken links in our documentation.

The vast majority (all?) were in the timer module. That was one of the first
module I documented with links and for some reason I thought that the ".html"
suffix was not needed when linking to things in the same module.

Where did I get that idea from, we'll never know.