write_dma: change API to consume buffer.
We take a reference to the buffer, but if we cancel the operation
it may be that the kernel could be writing to it. To avoid that we
take advantage of the fact that we keep the Source alive through the
operation and take ownership of the buffer during the write.