enhance local channel receiver with a recv function
Right now the local channel receiver can only be consumed through a stream.

The stream interface itself is convenient, as it can be used to fetch a single
element with next().

However, next(), try_next() and friends in futures_lite take a mutable reference
to self.

This is not very convenient when we are dealing with immutable references.
Because the receiver implementation is already entirely done with interior
mutability, we can easily expose a recv() interface.

Although in some situations we could put the receiver itself in a Cell
or Refcell, this is really not convenient:

* It is a rule of safety (which we should soon enforce with clippy), that
  you must never .await with a borrow held. Because everything in the same
  executor is executing in a single CPU, it is always safe to borrow() or
  borrow_mut() -> unless it crosses an .await point. As the result of
  .next() has to be .awaited, that breaks.

* A cell won't work well since the LocalReceiver is not Copy, and also not
  long lived: if we were to .take() it from the Cell (which is apparently
  legal even for non-Copy types), we'd have to leave the Cell empty while
  we are consuming.

The good thing about the Deque implementation which we had was that
pop_front() was &self, not &mut self. In fact as I try to change my
application to use channels instead of the Deque, I hit this one particular
issue every time.